pyproject.toml
[tool.poetry]
name = "python-craigslist-notifications"
version = "0.1.0"
description = ""
authors = ["whiteboarddev"]

[tool.poetry.dependencies]
python = "^3.7"
python-craigslist = "^1.0"
tinydb = "^3.15.1"
docopt = "^0.6.2"
schedule = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
